# open-api/tabliodb-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/snapshots": {
      "post": {
        "operationId": "createSnapshot",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/snapshots/diagram/{diagramId}": {
      "get": {
        "operationId": "getDiagramSnapshots",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "diagramId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotListResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/snapshots/{fromSnapshotId}/diff/{toSnapshotId}": {
      "get": {
        "operationId": "getSnapshotDiff",
        "parameters": [
          {
            "name": "toSnapshotId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromSnapshotId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotDiffResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/snapshots/{snapshotId}/restore": {
      "post": {
        "operationId": "restoreSnapshot",
        "parameters": [
          {
            "name": "snapshotId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "snapshots"
        ]
      }
    },
    "/teams": {
      "get": {
        "operationId": "getTeams",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamListResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "teams"
        ]
      },
      "post": {
        "operationId": "createTeam",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "teams"
        ]
      }
    },
    "/teams/{teamId}": {
      "patch": {
        "operationId": "updateTeam",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "teams"
        ]
      },
      "delete": {
        "operationId": "archiveTeam",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamArchiveResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "teams"
        ]
      }
    },
    "/teams/{teamId}/members": {
      "get": {
        "operationId": "getTeamMembers",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberListResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "teams"
        ]
      },
      "post": {
        "operationId": "addTeamMember",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamMemberCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "teams"
        ]
      }
    },
    "/teams/{teamId}/members/{userId}": {
      "delete": {
        "operationId": "removeTeamMember",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberRemoveResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "teams"
        ]
      }
    },
    "/teams/{teamId}/projects": {
      "get": {
        "operationId": "getTeamProjectAccesses",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamProjectAccessListResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "teams"
        ]
      },
      "post": {
        "operationId": "upsertTeamProjectAccess",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamProjectAccessUpsertDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamProjectAccessDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "teams"
        ]
      }
    },
    "/teams/{teamId}/projects/{projectId}": {
      "delete": {
        "operationId": "removeTeamProjectAccess",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamProjectAccessRemoveResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "",
        "tags": [
          "teams"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tabliodb API",
    "description": "Collaborative database schema designer API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tabliodb_access_token"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "SnapshotListResponseDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SnapshotResponseDto_Output"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "items",
          "nextCursor",
          "totalCount"
        ],
        "additionalProperties": false
      },
...
      "SnapshotDiffResponseDto_Output": {
        "type": "object",
        "properties": {
          "fromSnapshot": {
            "$ref": "#/components/schemas/SnapshotReferenceDto_Output"
          },
          "toSnapshot": {
            "$ref": "#/components/schemas/SnapshotReferenceDto_Output"
          },
          "migrationSql": {
            "$ref": "#/components/schemas/SnapshotMigrationSqlDto_Output"
          },
          "tables": {
            "$ref": "#/components/schemas/SnapshotTableChangeSummaryDto_Output"
          },
          "columns": {
            "$ref": "#/components/schemas/SnapshotEntityChangeSummaryDto_Output"
          },
          "relationships": {
            "$ref": "#/components/schemas/SnapshotEntityChangeSummaryDto_Output"
          },
          "indexes": {
            "$ref": "#/components/schemas/SnapshotEntityChangeSummaryDto_Output"
          },
          "enums": {
            "$ref": "#/components/schemas/SnapshotEntityChangeSummaryDto_Output"
          },
          "checks": {
            "$ref": "#/components/schemas/SnapshotEntityChangeSummaryDto_Output"
          },
          "notes": {
            "$ref": "#/components/schemas/SnapshotEntityChangeSummaryDto_Output"
          },
          "groups": {
            "$ref": "#/components/schemas/SnapshotEntityChangeSummaryDto_Output"
          },
          "dialectChanged": {
            "type": "boolean"
          },
          "metadataChanged": {
            "type": "boolean"
          },
          "schemaVersionChanged": {
            "type": "boolean"
          }
        },
        "required": [
          "fromSnapshot",
          "toSnapshot",
          "migrationSql",
          "tables",
          "columns",
          "relationships",
          "indexes",
          "enums",
          "checks",
          "notes",
          "groups",
          "dialectChanged",
          "metadataChanged",
          "schemaVersionChanged"
        ],
        "additionalProperties": false
      },
      "TeamResponseDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "memberCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "projectAccessCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "id",
          "organizationId",
          "name",
          "slug",
          "description",
          "memberCount",
          "projectAccessCount",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "TeamListResponseDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamResponseDto_Output"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "items",
          "nextCursor",
          "totalCount"
        ],
        "additionalProperties": false
      },
      "TeamCreateDto": {
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "organizationId",
          "name"
        ]
      },
      "TeamUpdateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "TeamArchiveResponseDto_Output": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          }
        },
        "required": [
          "successful"
        ],
        "additionalProperties": false
      },
      "TeamMemberDto_Output": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "name": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "cursorColor": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "userId",
          "email",
          "name",
          "avatarUrl",
          "cursorColor",
          "createdAt"
        ],
        "additionalProperties": false
      },
      "TeamMemberListResponseDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamMemberDto_Output"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "items",
          "nextCursor",
          "totalCount"
        ],
        "additionalProperties": false
      },
      "TeamMemberCreateDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          }
        },
        "required": [
          "email"
        ]
      },
      "TeamMemberRemoveResponseDto_Output": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          }
        },
        "required": [
          "successful"
        ],
        "additionalProperties": false
      },
...
      "TeamProjectAccessListResponseDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamProjectAccessDto_Output"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "items",
          "nextCursor",
          "totalCount"
        ],
        "additionalProperties": false
      },
      "TeamProjectAccessUpsertDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "role": {
            "default": "viewer",
            "type": "string",
            "enum": [
              "editor",
              "commenter",
              "viewer"
            ]
          }
        },
        "required": [
          "projectId"
        ]
      },
      "TeamProjectAccessRemoveResponseDto_Output": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          }
        },
        "required": [
          "successful"
        ],
        "additionalProperties": false
      },
...
    }
  }
}